+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+**TBD/TBD/TBD**
+
+  - Type-hinting: Fix mypy errors in vivarium/component.py
+
 **3.2.4 - 12/03/24**
 
   - Fix type hints for pandas groupby objects